% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{add_attachment}
\alias{add_attachment}
\title{Add an attachment to a post}
\usage{
add_attachment(post, file, attachment_name = "My File",
  host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{post}{Post object to attach to (message or comment)}

\item{file}{character, Path of file to attach}

\item{attachment_name}{character, Name of attachment seen in post, Default: 'My File'}

\item{host}{character, URL of the team, Default: Sys.getenv('BASECAMP_HOST')}

\item{token}{character, Basecamp Classic API token , Default: Sys.getenv("BASECAMP_TOKEN")}
}
\value{
Post with the attachment embedded in it.
}
\description{
Add an attachment object to a post
}
