% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{add_recipients}
\alias{add_recipients}
\title{Add Recipients to a Message}
\usage{
add_recipients(post, recipients)
}
\arguments{
\item{post}{post containing message to add recipients}

\item{recipients}{xml created by create_recipients}
}
\value{
character
}
\description{
Add recipients to a message
}
