% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{basecamp_project}
\alias{basecamp_project}
\alias{basecamp_projects}
\alias{basecamp_attachment}
\title{FUNCTION_TITLE}
\usage{
basecamp_project(project, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_projects(host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_attachment(project, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{project}{PARAM_DESCRIPTION}

\item{host}{PARAM_DESCRIPTION}

\item{token}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
