% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{basecamp_download}
\alias{basecamp_download}
\title{FUNCTION_TITLE}
\usage{
basecamp_download(url, destdir = getwd(), destfile = basename(url),
  token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{url}{PARAM_DESCRIPTION}

\item{destdir}{PARAM_DESCRIPTION, Default: getwd()}

\item{destfile}{PARAM_DESCRIPTION, Default: basename(url)}

\item{token}{PARAM_DESCRIPTION, Default: Sys.getenv("BASECAMP_TOKEN")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
