% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{create_notify}
\alias{create_notify}
\title{Create Notification}
\usage{
create_notify(person)
}
\arguments{
\item{person}{Name or ID of a person in Basecamp}
}
\value{
character
}
\description{
Create a notification node for a message object
}
