% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R, R/messages.R
\name{edit_title}
\alias{edit_title}
\alias{edit_body}
\alias{edit_category_id}
\alias{edit_milestone_id}
\alias{private_message}
\title{Edit a Basecamp Post}
\usage{
edit_title(object, value)

edit_body(object, value)

edit_category_id(object, value)

edit_milestone_id(object, value)

private_message(object, value)
}
\arguments{
\item{object}{post object}

\item{value}{new value to replace with}
}
\value{
updated post object
}
\description{
Edit post object fields
}
\details{
When editing the body basecamp is expecting HTML.

To make it simpler to write you can write a simple markdown text and
then convert it to html using \link[markdown:markdownToHTML]{markdownToHTML} :\preformatted{ html_text <- markdown::markdownToHTML(text = md_text,fragment.only = TRUE)
}

x will contain the html equivalent of value you can pass into \link[basecamper:edit_body]{edit_body}\preformatted{ new_message\%>\%
   edit_body(html_text)
}
}
