% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get.R
\name{find_person}
\alias{find_person}
\title{Find a person}
\usage{
find_person(pattern, ...)
}
\arguments{
\item{pattern}{character, pattern to search for}

\item{...}{arguments to pass to grepl}
}
\value{
named numeric vector
}
\description{
Find a person in Basecamp
}
\details{
If the index of people is not loaded in the package yet, it will be
 queried and stored in an internal package environment.
}
