% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{message_to_html}
\alias{message_to_html}
\title{FUNCTION_TITLE}
\usage{
message_to_html(x, index = 1, file = "")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION, Default: ''}

\item{head}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){

 project_messages <- basecamper::basecamp_messages('project','14542533')

 x <- message_to_html(project_messages)

 view_messages(x)

 }
}
}
