% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{notify_person}
\alias{notify_person}
\title{Notify a Person}
\usage{
notify_person(post, person)
}
\arguments{
\item{post}{post containing message to append notification to}

\item{person}{Name or ID of a person in Basecamp}
}
\value{
character
}
\description{
Create a notification node for a message object
}
\details{
DETAILS
}
