% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{post_comment}
\alias{post_comment}
\title{Post a comment}
\usage{
post_comment(comment, token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{comment}{object created by \link[basecamper:create_comment]{new_comment}}

\item{token}{character, Basecamp Classic API token , Default: Sys.getenv("BASECAMP_TOKEN")}
}
\value{
\link[httr:response]{response}
}
\description{
Post a comment to a Basecamp thread
}
