% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{post_message}
\alias{post_message}
\title{Post a message}
\usage{
post_message(message, token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{message}{object created by \link[basecamper:create_message]{new_message}}

\item{token}{character, Basecamp Classic API token , Default: Sys.getenv("BASECAMP_TOKEN")}
}
\value{
\link[httr:response]{response}
}
\description{
Post a message to Basecamp
}
