% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.R
\name{unzip2}
\alias{unzip2}
\title{FUNCTION_TITLE}
\usage{
unzip2(directory, file)
}
\arguments{
\item{directory}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[utils]{head}}
}
