% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{whoami}
\alias{whoami}
\title{whoami function for Basecamp}
\usage{
whoami(host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{host}{character, URL of the team, Default: Sys.getenv('BASECAMP_HOST')}

\item{token}{character, Basecamp Classic API token , Default: Sys.getenv("BASECAMP_TOKEN")}
}
\value{
xml_document
}
\description{
Query basic personal information for Basecamp account
}
