% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{add_attachment}
\alias{add_attachment}
\title{FUNCTION_TITLE}
\usage{
add_attachment(post, file, attachment_name = "My File",
  host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{post}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION}

\item{attachment_name}{PARAM_DESCRIPTION, Default: 'My File'}

\item{host}{PARAM_DESCRIPTION, Default: Sys.getenv("BASECAMP_HOST")}

\item{token}{PARAM_DESCRIPTION, Default: Sys.getenv("BASECAMP_TOKEN")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
