% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{basecamp_projects}
\alias{basecamp_projects}
\alias{basecamp_project}
\alias{basecamp_attachments}
\title{Basecamp Classic API wrappers}
\usage{
basecamp_projects(host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_project(project, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_attachments(project, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{host}{character, URL of the team, Default: Sys.getenv('BASECAMP_HOST')}

\item{token}{character, Basecamp Classic API token , Default: Sys.getenv("BASECAMP_TOKEN")}

\item{project}{character, Basecamp project id}
}
\value{
xml_document
}
\description{
Query projects, project and attachments in a project
}
\details{
A host url is from the following template: https://\{teamname\}.basecamphq.com
}
\examples{
\dontrun{
if(interactive()){

 (projects <- basecamp_projects())

 (summary_projects <- summary(projects))

 basecamp_project(summary_projects$id[1])

 (project_attachments <- basecamp_attachments(summary_projects$id[1]))

 summary(project_attachments)

 }
}
}
