% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{basecamp_download}
\alias{basecamp_download}
\title{Download file from Basecamp}
\usage{
basecamp_download(url, destdir = getwd(), destfile = basename(url),
  token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{url}{character, URL location of the file}

\item{destdir}{character, directory to save the file on local disk, Default: getwd()}

\item{destfile}{character, name of file to save on local disk, Default: basename(url)}

\item{token}{character, Basecamp Classic API token , Default: Sys.getenv("BASECAMP_TOKEN")}
}
\value{
a resoponse object
}
\description{
Download a file from Basecamp Classic
}
