% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{edit_title}
\alias{edit_title}
\alias{edit_body}
\alias{edit_category_id}
\alias{edit_milestone_id}
\alias{edit_privacy}
\title{Edit a Basecamp Post}
\usage{
edit_title(object, value)

edit_body(object, value)

edit_category_id(object, value)

edit_milestone_id(object, value)

edit_privacy(object, value)
}
\arguments{
\item{object}{post object}

\item{value}{new value to replace with}
}
\value{
post object
}
\description{
Edit post object fields
}
