% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{find_person}
\alias{find_person}
\title{Find a person}
\usage{
find_person(..., scope = "people", id = NULL,
  host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{...}{character, patterns to search for}

\item{scope}{character, scope of the request}

\item{id}{character, id relevant to the query}

\item{host}{character, URL of the team, Default: Sys.getenv('BASECAMP_HOST')}

\item{token}{character, Basecamp Classic API token , Default: Sys.getenv("BASECAMP_TOKEN")}
}
\value{
named numeric vector
}
\description{
Find a person in Basecamp
}
\details{
If the index of people is not loaded in the package yet, it will be
queried and stored in an internal package environment.
}
