% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get.R, R/attachments.R, R/comments.R,
%   R/messages.R, R/people.R
\name{basecamp_projects}
\alias{basecamp_projects}
\alias{basecamp_project}
\alias{basecamp_account}
\alias{basecamp_companies}
\alias{basecamp_categories}
\alias{basecamp_messages}
\alias{basecamp_attachments}
\alias{create_comment}
\alias{edit_comment}
\alias{delete_comment}
\alias{create_message}
\alias{basecamp_people}
\title{Basecamp Classic GET API wrappers}
\usage{
basecamp_projects(host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_project(id, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_account(host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_companies(scope = c("companies", "project", "company"),
  id = NULL, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_categories(scope = c("category", "project"), id = NULL,
  host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_messages(scope = c("message", "project"), id = NULL,
  host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_attachments(id, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

create_comment(scope = c("posts", "milestones", "todo_items"),
  id = NULL, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

edit_comment(host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

delete_comment(id = NULL, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

create_message(id = NULL, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))

basecamp_people(scope = c("person", "project", "company", "me",
  "people"), id = NULL, host = Sys.getenv("BASECAMP_HOST"),
  token = Sys.getenv("BASECAMP_TOKEN"))
}
\arguments{
\item{host}{character, URL of the team, Default: Sys.getenv('BASECAMP_HOST')}

\item{token}{character, Basecamp Classic API token , Default: Sys.getenv("BASECAMP_TOKEN")}

\item{id}{character, id relevant to the query}
}
\value{
xml_document
}
\description{
Query companies, people, projects, and attachments visible to the requesting user
}
\details{
A host url is from the following template: https://\{teamname\}.basecamphq.com
}
\examples{
\dontrun{
if(interactive()){

 (projects <- basecamp_projects())

 (summary_projects <- summary(projects))

 basecamp_project(summary_projects$id[1])

 (project_attachments <- basecamp_attachments(summary_projects$id[1]))

 summary(project_attachments)

 }
}
}
