% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{message_to_html}
\alias{message_to_html}
\title{Convert message content to HTML}
\usage{
message_to_html(message, index = 1, file = "")
}
\arguments{
\item{message}{message object returned by \link[basecamper:basecamp_message]{basecamper}}

\item{index}{numeric, indicies of comments to query (1 is earliest), Default: 1}

\item{file}{character, path to save the HTML output, Default: ''}
}
\value{
filepath to the html created
}
\description{
Uses the message ID in the message object to query the comments
associated with it and
}
\details{
if file is '' then a tmepfile will be written created and written to.
}
\examples{
\dontrun{
if(interactive()){

 project_messages <- basecamper::basecamp_messages('project','14542533')

 x <- message_to_html(project_messages)

 view_messages(x)

 }
}
}
