% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.R
\name{unzip2}
\alias{unzip2}
\title{Unzip large files}
\usage{
unzip2(directory, file)
}
\arguments{
\item{directory}{character, path of directory to unzip the file}

\item{file}{character, local path of the file to unzip.}
}
\description{
CLI call to unzip that is more robust for large files, which
are truncated by \code{\link[utils]{unzip}} if larger than 4gb.
}
\seealso{
\code{\link[utils]{unzip}}
}
