% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{view_messages}
\alias{view_messages}
\title{View Messages}
\usage{
view_messages(file, viewer = getOption("viewer"))
}
\arguments{
\item{x}{View Message in Viewer}

\item{head}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
